+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- CLI: add the `-se` argument to `ewoksjob worker` to define SLURM job environment variables.
+
 ## [1.4.1] - 2026-02-17
 
 ### Changed